Moved Response.hpp to responses subdir: responses/Response.hpp
