+++ includes/responses/BadStatusResponse.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <Response.hpp>
+#include <responses/Response.hpp>
 
 namespace Webserver
 {
