+++ srcs/responses/Response.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 #include <Logger.hpp>
-#include <Response.hpp>
+#include <responses/Response.hpp>
 #include <Utility.hpp>
 #include <Exception.hpp>
 #include <config/GlobalConfig.hpp>
